component correction
