+++ client/src/components/sidebar/index.tsx
@@ -11,9 +11,9 @@ export default function Sidebar() {
   const router = useRouter();
   
   return (
-    <div className="sidebar flex flex-col h-full items-center justify-between bg-backgroundSidebar w-64">
+    <div className="flex flex-col h-full items-center justify-between bg-backgroundSidebar w-64 p-6">
 
-      <div className="sidebar_UpSide flex flex-col items-center gap-4 py-7">
+      <div className="flex flex-col items-center gap-4 py-7">
         <Image src={logoCITiWhite} alt="Logo" className="itens-center w-30 h-auto mb-4"/>
 
         <Button onClick={() => router.push("/dashboard")} className="flex items-center justify-start w-56 h-11 px-4 bg-white bg-opacity-40 [&_svg]:size-8 hover:bg-black/20 hover:shadow-xl">
@@ -27,7 +27,7 @@ export default function Sidebar() {
         </Button>
       </div>
 
-      <div className="sidebar_footer flex flex-row items-center gap-2 p-4">
+      <div className="flex flex-row items-center gap-2">
         <p className="text-white text-sm font-barlow">
           Made with <strong>&lt; &#x0002F; &gt;</strong> and{" "}<strong>&hearts;</strong> by{" "}
         </p>
